requirements.txt
aiogram>=3.0,<4.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
schedule==1.2.0
asyncio==3.4.3
loguru==0.7.2
openai>=1.3.7
beautifulsoup4==4.12.2
matplotlib==3.7.2
pillow==10.0.0
pytz==2023.3